requirements.txt
absl-py==2.2.2
diff_gaussian_rasterization_depth_alpha==0.0.0
filelock==3.18.0
freetype-py==2.5.1
fsspec==2025.3.2
fused==0.0.0
glfw==2.9.0
grpcio==1.71.0
imageio==2.37.0
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.0
lazy_loader==0.4
libigl==2.5.1
Markdown==3.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
plyfile==1.1
portalocker==3.1.1
protobuf==6.30.2
pyglet==2.1.6
PyOpenGL==3.1.0
pyrender==0.1.45
pytorch-fid==0.3.0
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@0a59450f0ebbe12d9a8db3de937814932517633b
PyYAML==6.0.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
screeninfo==0.8.1
setuptools==80.3.1
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
trimesh==4.6.8
triton==3.3.0
typing_extensions==4.13.2
upfirdn2d==0.0.0
Werkzeug==3.1.3
wheel==0.45.1
